Add quickreply support to discord handler
Even though discord doesn't currently support quick replies, we can
still pass this over to the proxy so that we don't need to change
courier when it does. In the meantime, the proxy can do a partial
attempt to support single emoji quick replies in the form of reactions,
which is perfectly servicable for some purposes.